# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Greenhouse Backend API
  description: 농장 관리 시스템 Backend API
  version: 0.0.1
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/business-partners/{partnerId}/settlement-settings:
    get:
      tags:
      - partner-settlement-settings-controller
      operationId: get
      parameters:
      - name: partnerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePartnerSettlementSettingsResponse'
    put:
      tags:
      - partner-settlement-settings-controller
      operationId: update_1
      parameters:
      - name: partnerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartnerSettlementSettingsRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePartnerSettlementSettingsResponse'
...
  /api/auction-settlements/rebuild:
    post:
      tags:
      - auction-settlement-controller
      operationId: rebuild
      parameters:
      - name: auctionHouseId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: auctionDate
        in: query
        required: true
        schema:
          type: string
          format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAuctionSettlementResponse'
...
  /api/auction-settlements:
    get:
      tags:
      - auction-settlement-controller
      description: 호환용 목록. 최신 500건만 반환합니다. 운영 목록은 /page, 전체 합계는 /summary를 사용합니다.
      operationId: getSettlements
      parameters:
      - name: auctionHouseId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - CREATED
          - PAYMENT_WAITING
          - PARTIALLY_PAID
          - PAID
          - AMOUNT_MISMATCH
          - REVIEW_REQUIRED
          - CANCELLED
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseListAuctionSettlementResponse'
      deprecated: true
  /api/auction-settlements/{settlementId}:
    get:
      tags:
      - auction-settlement-controller
      operationId: getSettlement
      parameters:
      - name: settlementId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAuctionSettlementResponse'
  /api/auction-settlements/summary:
    get:
      tags:
      - auction-settlement-controller
      description: 조회 조건에 해당하는 전체 정산의 합계. 페이지 크기와 무관하게 집계합니다.
      operationId: getAuctionSettlementSummary
      parameters:
      - name: auctionHouseId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - CREATED
          - PAYMENT_WAITING
          - PARTIALLY_PAID
          - PAID
          - AMOUNT_MISMATCH
          - REVIEW_REQUIRED
          - CANCELLED
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAuctionSettlementSummaryResponse'
  /api/auction-settlements/page:
    get:
      tags:
      - auction-settlement-controller
      description: 정산 요약의 서버 페이지 목록. page는 0 이상, size는 1~100으로 보정합니다. 상세 행은 단건 API에서 조회합니다.
      operationId: getAuctionSettlementPage
      parameters:
      - name: auctionHouseId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - CREATED
          - PAYMENT_WAITING
          - PARTIALLY_PAID
          - PAID
          - AMOUNT_MISMATCH
          - REVIEW_REQUIRED
          - CANCELLED
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageResponseAuctionSettlementListItemResponse'
...
components:
  schemas:
...
    PartnerSettlementSettingsRequest:
      type: object
      properties:
        allowPrepayment:
          type: boolean
        amountTolerance:
          type: integer
          format: int64
          minimum: 0
        autoMatchEnabled:
          type: boolean
        autoSettleEnabled:
          type: boolean
        creditAutoApplyEnabled:
          type: boolean
        depositorAliases:
          type: array
          items:
            type: string
            maxLength: 100
            minLength: 0
        memo:
          type: string
          maxLength: 1000
          minLength: 0
        paymentDayMode:
          type: string
          enum:
          - CALENDAR_DAY
          - BUSINESS_DAY
        paymentDelayDays:
          type: integer
          format: int32
          minimum: 0
        ruleJson:
          type: object
          additionalProperties: {}
        settlementUnit:
          type: string
          enum:
          - SALES_SLIP
          - MONTHLY_BATCH
          - AUCTION_DATE
      required:
      - amountTolerance
      - depositorAliases
      - paymentDayMode
      - paymentDelayDays
      - settlementUnit
    ApiResponsePartnerSettlementSettingsResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/PartnerSettlementSettingsResponse'
        message:
          type: string
    PartnerSettlementSettingsResponse:
      type: object
      properties:
        allowPrepayment:
          type: boolean
        amountTolerance:
          type: integer
          format: int64
        autoMatchEnabled:
          type: boolean
        autoSettleEnabled:
          type: boolean
        creditAutoApplyEnabled:
          type: boolean
        depositorAliases:
          type: array
          items:
            type: string
        id:
          type: integer
          format: int64
        memo:
          type: string
        partnerId:
          type: integer
          format: int64
        paymentDayMode:
          type: string
          enum:
          - CALENDAR_DAY
          - BUSINESS_DAY
        paymentDelayDays:
          type: integer
          format: int32
        ruleJson:
          type: object
          additionalProperties: {}
        settlementUnit:
          type: string
          enum:
          - SALES_SLIP
          - MONTHLY_BATCH
          - AUCTION_DATE
...
    ApiResponseAuctionSettlementResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/AuctionSettlementResponse'
        message:
          type: string
    AuctionSettlementLineResponse:
      type: object
      properties:
        amount:
          type: integer
          format: int64
        auctionResultLineId:
          type: integer
          format: int64
        auctionShipmentLotId:
          type: integer
          format: int64
        id:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
        shipmentDate:
          type: string
          format: date
        shipmentGrade:
          type: string
        status:
          type: string
          enum:
          - UNPAID
          - PAID
          - PARTIALLY_PAID
          - EXCLUDED
          - REVIEW_REQUIRED
        unitPrice:
          type: integer
          format: int32
        varietyName:
          type: string
    AuctionSettlementResponse:
      type: object
      properties:
        auctionDate:
          type: string
          format: date
        auctionHouseId:
          type: integer
          format: int64
        auctionHouseName:
          type: string
        confirmedAt:
          type: string
          format: date-time
        confirmedBy:
          type: string
        deductionAmount:
          type: integer
          format: int64
        expectedDepositAmount:
          type: integer
          format: int64
        expectedPaymentDate:
          type: string
          format: date
        feeAmount:
          type: integer
          format: int64
        grossAmount:
          type: integer
          format: int64
        id:
          type: integer
          format: int64
        lines:
          type: array
          items:
            $ref: '#/components/schemas/AuctionSettlementLineResponse'
        memo:
          type: string
        paidAmount:
          type: integer
          format: int64
        remainingAmount:
          type: integer
          format: int64
        resultReceivedAt:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - CREATED
          - PAYMENT_WAITING
          - PARTIALLY_PAID
          - PAID
          - AMOUNT_MISMATCH
          - REVIEW_REQUIRED
          - CANCELLED
...
    ApiResponseListAuctionSettlementResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AuctionSettlementResponse'
        message:
          type: string
    ApiResponseAuctionSettlementSummaryResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/AuctionSettlementSummaryResponse'
        message:
          type: string
    AuctionSettlementSummaryResponse:
      type: object
      properties:
        expectedDepositAmount:
          type: integer
          format: int64
        remainingAmount:
          type: integer
          format: int64
    ApiResponsePageResponseAuctionSettlementListItemResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/PageResponseAuctionSettlementListItemResponse'
        message:
          type: string
    AuctionSettlementListItemResponse:
      type: object
      properties:
        auctionDate:
          type: string
          format: date
        auctionHouseId:
          type: integer
          format: int64
        auctionHouseName:
          type: string
        expectedDepositAmount:
          type: integer
          format: int64
        grossAmount:
          type: integer
          format: int64
        id:
          type: integer
          format: int64
        remainingAmount:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - CREATED
          - PAYMENT_WAITING
          - PARTIALLY_PAID
          - PAID
          - AMOUNT_MISMATCH
          - REVIEW_REQUIRED
          - CANCELLED
    PageResponseAuctionSettlementListItemResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AuctionSettlementListItemResponse'
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...